pyproject.toml
[tool.ruff]
output-format = "full"
lint.select = ["ALL"]
line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".eggs",
    "*.egg",
    "dist",
    "tests/fixtures/**",
    "tests/**/snapshots/**",
    "alembic",
    "airich",
]
lint.ignore = [
    "D100",
    "B008",
    "D104",
    "Q000",
    "S101",
    "PT016",
    "ANN101",
    "ANN102",
    "N805",
    "UP037",
    "PLC0414",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
ignore-decorators = ["typing.overload"]

[tool.ruff.lint.mccabe]
max-complexity = 11

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query"]

[tool.ruff.lint.extend-per-file-ignores]
"typings/*" = ["ANN401"]
"__init__.py" = ["F401"]
"manage.py" = ["E402"]
"tests/*" = ["D103"]


[tool.black]
line-length = 100
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3


[tool.coverage]
[tool.coverage.run]
source = ["sqlrepo"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "enum.Enum",
    "(Protocol):",
    "(typing.Protocol):",
    "pragma: no cover",
    "pragma: no coverage",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@overload",
]

[tool.pytest]
testpath = "tests"


[tool.bandit]
exclude = ['tests']

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.5.5",
    "vulture>=2.11",
    "pytest>=8.1.1",
    "black>=24.3.0",
    "bandit>=1.7.8",
    "coverage>=7.4.4",
    "pytest-cov>=4.1.0",
    "isort>=5.13.2",
    "pyright>=1.1.355",
    "freezegun>=1.4.0",
    "mimesis>=15.1.0",
    "ipython>=8.22.2",
    "sqlalchemy-utils>=0.41.2",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "pytest-asyncio>=0.23.6",
    "httpx>=0.27.0",
    "pyment>=0.3.3",
    "ipython>=8.19.0",
]


[project]
name = "sqlalchemy_profiler"
version = "1.0.0"
description = "SQLAlchemy profiler classes and extentions for other packages."
authors = [{ name = "Dmitriy Lunev", email = "dima.lunev14@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = ["sqlalchemy>=2.0", "python-dev-utils==7.0.1"]

[project.optional-dependencies]
fastapi = ["fastapi"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
